Split serde helpers into separate files
